Add support for UVC RGB24 format, and dithered YUY2 (default)

+++ firmware/src/usb/device_video/usb_video_device.hh
@@ -13,8 +13,10 @@ public:
 	void start();
 	void stop();
 
-	// Set the pointer to the YUY2 shadow framebuffer (320x240, pre-converted on A7)
-	static void set_framebuffer(uint16_t *fb);
+	// Set the pointer to the shadow framebuffer (pre-converted on A7).
+	// Contents are either YUY2 (2 bytes/pixel) or BGR24 (3 bytes/pixel),
+	// selected at compile time via USE_UVC_FORMAT_BGR3.
+	static void set_framebuffer(uint8_t *fb);
 
 	static int8_t Video_Itf_Init();
 	static int8_t Video_Itf_DeInit();
@@ -24,7 +26,7 @@ public:
 private:
 	USBD_HandleTypeDef *pdev;
 
-	static inline uint16_t *shadow_fb = nullptr;
+	static inline uint8_t *shadow_fb = nullptr;
 	static inline uint32_t pixel_offset = 0;
 	static inline uint32_t packet_index = 0;
 };
